feat: Update device data models to use Double for CPM and Batvolt
- Changed CPM field type from Integer to Double in EnvironmentDeviceDataResponse and RadiationDeviceDataResponse.
- Updated CPM field type from Integer to Double in EnvironmentDeviceData and RadiationDeviceData entities.
- Added description field to Company entity with validation.
- Introduced new fields (manufacturer, model, serialNumber, installDate) in Device entity.
- Implemented data initialization logic in DataInitializer for default company, users, and devices.
- Created DeviceDataReceiverController to handle incoming device data for radiation and environment devices.
- Added request DTOs for receiving device data (RadiationDataReceiveRequest, EnvironmentDataReceiveRequest).
- Added response DTO for device data reception (DeviceDataReceiveResponse).
- Created enums for UserRole and UserStatus to manage user roles and statuses.
- Updated UserRepository and UserService to use new enums for user role and status.
- Enhanced SecurityConfig to allow unauthenticated access to device data APIs.

+++ backend/src/main/java/com/cdutetc/ems/dto/response/EnvironmentDeviceDataResponse.java
@@ -21,7 +21,7 @@ public class EnvironmentDeviceDataResponse {
     private String deviceCode;
     private String rawData;
     private Integer src;
-    private Integer CPM;
+    private Double CPM;
     private Double temperature;
     private Double wetness;
     private Double windspeed;
